requirements.txt
appdirs==1.4.4
audioread==3.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
decorator==5.1.1
ffmpeg-python==0.2.0
future==0.18.3
idna==3.4
joblib==1.2.0
lazy_loader==0.2
librosa==0.10.0.post2
llvmlite==0.40.0
msgpack==1.0.5
numba==0.57.0
numpy==1.24.3
packaging==23.1
pooch==1.6.0
pycparser==2.21
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
soundfile==0.12.1
soxr==0.3.5
threadpoolctl==3.1.0
typing_extensions==4.5.0
urllib3==2.0.2
